--- ImageFeed/Extensions/Date+Extensions.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension String {
-    func toDate() -> Date? {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return dateFormatter.date(from: self)
-    }
-}
