+++ ImageFeed/Models/ImagesListPhoto.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+struct Photo {
+    let id: String
+    let size: CGSize
+    let createdAt: Date?
+    let welcomeDescription: String?
+    let thumbImageURL: String
+    let largeImageURL: String
+    var isLiked: Bool
+}
